기능 구현 완료(Hover freeze까지), context GroupSpacePage로 옮김
