+++ src/App.jsx
@@ -7,191 +7,36 @@ import GroupHomePage from "./pages/GroupHomePage";
 import NoGroupPage from "./pages/NoGroupPage";
 import { useAuthStatus } from "./hooks/useAuthStatus";
 import GroupSpacePage from "./Pages/GroupSpacePage";
-import { useState, useReducer, createContext, useRef } from "react";
-
-export const toCleanStateContext = createContext();
-export const toCleanDispatchContext = createContext();
-
-const personMockData = [
-    {
-        name: "A",
-        badgeId: 1,
-        email: "A@email.com",
-        pw: "1111",
-        cleanSensitivity: 50,
-    },
-    {
-        name: "B",
-        badgeId: 2,
-        email: "B@email.com",
-        pw: "2222",
-        cleanSensitivity: 80,
-    },
-    {
-        name: "C",
-        badgeId: 3,
-        email: "C@email.com",
-        pw: "333",
-        cleanSensitivity: 30,
-    },
-    {
-        name: "D",
-        badgeId: 4,
-        email: "D@email.com",
-        pw: "444",
-        cleanSensitivity: 20,
-    },
-];
-
-// 한번에 모든 정보를 담고, map으로 찾을 생각 (첨엔, group/person으로 나눴었는데, 짜피 target=group/personal 정보도 저장하니 굳이 싶어 구분없앰)
-const checkListMockData = [
-    {
-        target: "group",
-        id: 1,
-        name: "A",
-        badgeId: 1,
-        place: "거실",
-        toClean: "tv 닦기",
-        deadLine: "D-2",
-    },
-    {
-        target: "group",
-        id: 2,
-        name: "B",
-        badgeId: 2,
-        place: "화장실",
-        toClean: "거울 닦기",
-        deadLine: "D-day",
-    },
-    {
-        target: "person",
-        id: 3,
-        name: "A",
-        badgeId: 1,
-        place: "책상",
-        toClean: "책상 정리",
-        deadLine: "D-2",
-    },
-    {
-        target: "person",
-        id: 4,
-        name: "B",
-        badgeId: 2,
-        place: "침대",
-        toClean: "이불 개기",
-        deadLine: "D-day",
-    },
-];
-
-function reducer(data, action) {
-    switch (action.type) {
-        case "CREATE":
-            return [...data, action.data];
-        case "UPDATE":
-            return data.map((item) => {
-                if (String(item.id) === String(action.data.id)) {
-                    return action.data;
-                }
-                return item;
-            });
-        case "DELETE":
-            return data.filter((item) => String(item.id) !== String(action.id));
-        default:
-            return data;
-    }
-}
 
 function App() {
     const { isLoggedIn, hasGroup } = useAuthStatus();
-    const [checkListData, dispatch] = useReducer(reducer, checkListMockData);
-    const [personData, setPersonData] = useState(personMockData);
-    const idRef = useRef(5);
-
-    const onCreate = (target, name, badgeId, place, toClean, deadLine) => {
-        dispatch({
-            type: "CREATE",
-            data: {
-                target,
-                id: idRef.current++,
-                name,
-                badgeId,
-                place,
-                toClean,
-                deadLine,
-            },
-        });
-    };
-
-    const onUpdate = (target, id, name, badgeId, place, toClean, deadLine) => {
-        dispatch({
-            type: "UPDATE",
-            target,
-            data: {
-                target,
-                id,
-                name,
-                badgeId,
-                place,
-                toClean,
-                deadLine,
-            },
-        });
-    };
-
-    const onDelete = (target, id) => {
-        dispatch({
-            type: "DELETE",
-            target,
-            id,
-        });
-    };
 
     return (
         <>
-            <toCleanDispatchContext.Provider
-                value={{ onCreate, onUpdate, onDelete }}
-            >
-                <toCleanStateContext.Provider
-                    value={{ checkListData, personData }}
-                >
-                    <Routes>
-                        <Route path="/" element={<LoginPage />} />
-                        <Route path="/noGroup" element={<NoGroupPage />} />
-                        <Route
-                            path="/createGroup"
-                            element={<CreateGroupPage />}
-                        />
-                        <Route path="/tutorial" element={<TutorialPage />} />
-                        <Route
-                            path="/createSpace"
-                            element={<CreateSpacePage />}
-                        />
-                        <Route
-                            path="/groupSpace"
-                            element={<GroupSpacePage />}
-                        />
-                        <Route path="/groupHome" element={<GroupHomePage />} />
-                        <Route
-                            path="/redirect"
-                            element={
-                                isLoggedIn ? (
-                                    hasGroup ? (
-                                        <Navigate to="/createSpace" replace />
-                                    ) : (
-                                        <Navigate to="/createGroup" replace />
-                                    )
-                                ) : (
-                                    <Navigate to="/" replace />
-                                )
-                            }
-                        />
-                        <Route
-                            path="*"
-                            element={<div>잘못된 페이지입니다.</div>}
-                        />
-                    </Routes>
-                </toCleanStateContext.Provider>
-            </toCleanDispatchContext.Provider>
+            <Routes>
+                <Route path="/" element={<LoginPage />} />
+                <Route path="/noGroup" element={<NoGroupPage />} />
+                <Route path="/createGroup" element={<CreateGroupPage />} />
+                <Route path="/tutorial" element={<TutorialPage />} />
+                <Route path="/createSpace" element={<CreateSpacePage />} />
+                <Route path="/groupSpace" element={<GroupSpacePage />} />
+                <Route path="/groupHome" element={<GroupHomePage />} />
+                <Route
+                    path="/redirect"
+                    element={
+                        isLoggedIn ? (
+                            hasGroup ? (
+                                <Navigate to="/createSpace" replace />
+                            ) : (
+                                <Navigate to="/createGroup" replace />
+                            )
+                        ) : (
+                            <Navigate to="/" replace />
+                        )
+                    }
+                />
+                <Route path="*" element={<div>잘못된 페이지입니다.</div>} />
+            </Routes>
         </>
     );
 }
